Port subset ops module into rook
