+++ src/rook/utils/ops/__init__.py
@@ -3,6 +3,7 @@
 from .base import Operation
 from .average import Average, average_over_dims, average_shape, average_time
 from .regrid import Regrid, regrid
+from .subset import Subset, subset
 from .normalise import ResultSet, normalise
 
 __all__ = [
@@ -13,6 +14,8 @@ __all__ = [
 	"average_time",
 	"Regrid",
 	"regrid",
+	"Subset",
+	"subset",
 	"ResultSet",
 	"normalise",
 ]
